Clean-up of trading dates code.
- Made td an integer.

+++ crsp/make_trading_dates.sql
@@ -1,10 +1,10 @@
--- SELECT CURRENT_DATE + s.a AS dates 
+-- SELECT CURRENT_DATE + s.a AS dates
 --   FROM generate_series(0,14,7) as s(a);
 -- DROP TABLE IF EXISTS crsp.trading_dates CASCADE;
 DROP TABLE IF EXISTS crsp.trading_dates CASCADE;
 
 CREATE TABLE crsp.trading_dates AS
-SELECT date, rank() OVER (ORDER BY date) AS td
+SELECT date, (rank() OVER (ORDER BY date))::integer AS td
     FROM (SELECT DISTINCT DATE FROM crsp.dsi) AS a;
 
 CREATE INDEX ON crsp.trading_dates (date);
@@ -14,7 +14,7 @@ ANALYZE crsp.trading_dates;
 DROP VIEW IF EXISTS crsp.all_dates;
 
 CREATE VIEW crsp.all_dates AS
-SELECT anncdate::date 
+SELECT anncdate::date
 FROM generate_series((SELECT min(date) FROM crsp.trading_dates)::timestamp,
 		(SELECT max(date) FROM crsp.trading_dates), '1 day') AS anncdate;
 
@@ -24,7 +24,7 @@ CREATE TABLE crsp.anncdates AS
 SELECT a.anncdate, min(td) OVER w AS td, min(date) OVER w AS date
 FROM crsp.all_dates AS a
 LEFT JOIN crsp.trading_dates AS b
-ON b.date = a.anncdate 
+ON b.date = a.anncdate
 WINDOW w AS (ORDER BY anncdate ROWS BETWEEN CURRENT ROW AND 7 FOLLOWING)
 ORDER BY anncdate;
